[powerpc] deactivate tests combined.test and object.test on powerpc64 be
These tests fail on powerpc64 BE (only, not LE) and are thus intefering with
the running of 3 of the powerpc buildbots.  The author of the tests has been
notified and is working on fixing them but in the meantime I am disabling
them to get the bots working again.

See https://bugs.llvm.org//show_bug.cgi?id=33429


git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@305317 91177308-0d34-0410-b5e6-96231b3b80d8