[codeview] Ignore DBG_VALUEs when choosing a BB start source loc
When the first instruction of a basic block has no location (consider a
LEA materializing the address of an alloca for a call), we want to start
the line table for the block with the first valid source location in the
block.  We need to ignore DBG_VALUE instructions during this scan to get
decent line tables.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@309628 91177308-0d34-0410-b5e6-96231b3b80d8